+++ core/core_translations/fr.ts
@@ -2,189 +2,217 @@
 <!DOCTYPE TS>
 <TS version="2.1" language="fr" sourcelanguage="en">
 <context>
+    <name>MDate</name>
+    <message>
+        <location filename="../mdate.cpp" line="170"/>
+        <source>Winter</source>
+        <translation>Hiver</translation>
+    </message>
+    <message>
+        <location filename="../mdate.cpp" line="172"/>
+        <source>Spring</source>
+        <translation>Printemps</translation>
+    </message>
+    <message>
+        <location filename="../mdate.cpp" line="174"/>
+        <source>Summer</source>
+        <translation>Été</translation>
+    </message>
+    <message>
+        <location filename="../mdate.cpp" line="176"/>
+        <source>Fall</source>
+        <translation>Automne</translation>
+    </message>
+</context>
+<context>
     <name>Print</name>
     <message>
-        <location filename="../print.cpp" line="56"/>
+        <location filename="../print.cpp" line="67"/>
         <source>General crop plan</source>
         <translation>Plane de culture général</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="70"/>
-        <location filename="../print.cpp" line="112"/>
-        <location filename="../print.cpp" line="150"/>
-        <location filename="../print.cpp" line="189"/>
+        <location filename="../print.cpp" line="81"/>
+        <location filename="../print.cpp" line="123"/>
+        <location filename="../print.cpp" line="161"/>
+        <location filename="../print.cpp" line="200"/>
         <source>Crop</source>
         <translation>Espèce</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="71"/>
-        <location filename="../print.cpp" line="113"/>
-        <location filename="../print.cpp" line="151"/>
-        <location filename="../print.cpp" line="190"/>
+        <location filename="../print.cpp" line="82"/>
+        <location filename="../print.cpp" line="124"/>
+        <location filename="../print.cpp" line="162"/>
+        <location filename="../print.cpp" line="201"/>
         <source>Variety</source>
         <translation>Variété</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="72"/>
-        <location filename="../print.cpp" line="111"/>
-        <location filename="../print.cpp" line="149"/>
+        <location filename="../print.cpp" line="83"/>
+        <location filename="../print.cpp" line="122"/>
+        <location filename="../print.cpp" line="160"/>
         <source>S</source>
         <translation>S</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="73"/>
-        <location filename="../print.cpp" line="114"/>
-        <location filename="../print.cpp" line="188"/>
+        <location filename="../print.cpp" line="84"/>
+        <location filename="../print.cpp" line="125"/>
+        <location filename="../print.cpp" line="199"/>
         <source>TP</source>
         <translation>P</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="74"/>
-        <location filename="../print.cpp" line="152"/>
-        <location filename="../print.cpp" line="191"/>
+        <location filename="../print.cpp" line="85"/>
+        <location filename="../print.cpp" line="163"/>
+        <location filename="../print.cpp" line="202"/>
         <source>FH</source>
         <translation>DR</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="75"/>
+        <location filename="../print.cpp" line="86"/>
         <source>LH</source>
         <translation>FR</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="76"/>
-        <location filename="../print.cpp" line="153"/>
-        <location filename="../print.cpp" line="192"/>
+        <location filename="../print.cpp" line="87"/>
+        <location filename="../print.cpp" line="164"/>
+        <location filename="../print.cpp" line="203"/>
         <source>Length</source>
         <translation>Lgr.</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="77"/>
-        <location filename="../print.cpp" line="155"/>
-        <location filename="../print.cpp" line="194"/>
+        <location filename="../print.cpp" line="88"/>
+        <location filename="../print.cpp" line="166"/>
+        <location filename="../print.cpp" line="205"/>
         <source>Rows</source>
         <translation>Rangs</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="78"/>
+        <location filename="../print.cpp" line="89"/>
         <source>Spac.</source>
         <translation>Dist.</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="79"/>
-        <location filename="../print.cpp" line="157"/>
-        <location filename="../print.cpp" line="197"/>
-        <location filename="../print.cpp" line="226"/>
+        <location filename="../print.cpp" line="90"/>
+        <location filename="../print.cpp" line="168"/>
+        <location filename="../print.cpp" line="208"/>
+        <location filename="../print.cpp" line="237"/>
         <source>Locations</source>
         <translation>Emplacements</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="98"/>
+        <location filename="../print.cpp" line="109"/>
         <source>Greenhouse crop plan</source>
         <translation>Plan de culture pépinière</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="115"/>
-        <location filename="../print.cpp" line="195"/>
+        <location filename="../print.cpp" line="126"/>
+        <location filename="../print.cpp" line="206"/>
         <source># trays</source>
         <translation>Nb. plaques</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="116"/>
-        <location filename="../print.cpp" line="196"/>
+        <location filename="../print.cpp" line="127"/>
+        <location filename="../print.cpp" line="207"/>
         <source>Size</source>
         <translation>Taille</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="117"/>
+        <location filename="../print.cpp" line="128"/>
         <source>Seeds/hole</source>
         <translation>Graines/trou</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="118"/>
+        <location filename="../print.cpp" line="129"/>
         <source># seeds</source>
         <translation>Nb. graines</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="119"/>
+        <location filename="../print.cpp" line="130"/>
         <source>Seed weight</source>
         <translation>Poids</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="136"/>
+        <location filename="../print.cpp" line="147"/>
         <source>Field sowing crop plan</source>
         <translation>Plan de culture semis direct</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="154"/>
-        <location filename="../print.cpp" line="193"/>
+        <location filename="../print.cpp" line="165"/>
+        <location filename="../print.cpp" line="204"/>
         <source>Spacing</source>
         <translation>Dist.</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="156"/>
+        <location filename="../print.cpp" line="167"/>
         <source>Weight</source>
         <translation>Poids</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="174"/>
+        <location filename="../print.cpp" line="185"/>
         <source>Field transplanting crop plan</source>
         <translation>Plan de culture plantation</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="215"/>
+        <location filename="../print.cpp" line="226"/>
         <source>Task calendar</source>
         <translation>Calendrier des tâches</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="224"/>
+        <location filename="../print.cpp" line="235"/>
         <source>Date</source>
         <translation>Date</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="225"/>
+        <location filename="../print.cpp" line="236"/>
         <source>Planting</source>
         <translation>Série</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="227"/>
+        <location filename="../print.cpp" line="238"/>
         <source>Description</source>
         <translation>Description</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="228"/>
+        <location filename="../print.cpp" line="239"/>
         <source>Notes</source>
         <translation>Notes</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="425"/>
+        <location filename="../print.cpp" line="436"/>
         <source> W%1, %2 − %3</source>
         <translation> S%1, %2 − %3</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="481"/>
+        <location filename="../print.cpp" line="492"/>
         <source>%1 beds</source>
         <translation>%1 pl.</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="483"/>
+        <location filename="../print.cpp" line="494"/>
         <source>%1 bed m.</source>
         <translation>%1 m pl.</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="489"/>
         <location filename="../print.cpp" line="500"/>
+        <location filename="../print.cpp" line="511"/>
         <source>%1, %2 rows x %3 cm</source>
         <translation>%1, %2 rangs x %3 cm</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="494"/>
+        <location filename="../print.cpp" line="505"/>
         <source>%1 x %2, %3 seeds per hole</source>
         <translation>%1 x %2, %3 graines par trou</translation>
     </message>
     <message>
-        <location filename="../print.cpp" line="496"/>
+        <location filename="../print.cpp" line="507"/>
         <source>%1 x %2</source>
         <translation></translation>
+    </message>
+    <message>
+        <location filename="../print.cpp" line="600"/>
+        <source>Location</source>
+        <translation>Emplacement</translation>
     </message>
 </context>
 </TS>