+++ core/mdate.h
@@ -36,6 +36,7 @@ class CORESHARED_EXPORT MDate : public QObject
 public:
     explicit MDate(QObject *parent = nullptr);
 
+    static const QList<QList<int>> monthsOrder;
     Q_INVOKABLE static QDate dateFromWeekString(const QString &s);
     Q_INVOKABLE static QDate dateFromDateString(const QString &s);
     static QDate firstMondayOfYear(int year);
@@ -48,14 +49,18 @@ public:
     Q_INVOKABLE static int currentYear();
     Q_INVOKABLE static QString formatDate(const QDate &date, int currentYear,
                                           const QString &type = "", const bool showIndicator = true);
+
     Q_INVOKABLE static int season(const QDate &date);
     Q_INVOKABLE static int seasonYear(const QDate &date);
+    static QString seasonName(int season);
     static QPair<QDate, QDate> seasonDates(int season, int year);
     Q_INVOKABLE static QDate seasonBeginning(int season, int year);
+
     Q_INVOKABLE static qint64 daysTo(const QDate &from, const QDate &to) { return from.daysTo(to); }
     Q_INVOKABLE static QDate addDays(const QDate &date, qint64 days) { return date.addDays(days); }
     Q_INVOKABLE static QString dayName(const QDate &date);
     Q_INVOKABLE static QString monthName(int month);
+    Q_INVOKABLE static QString shortMonthName(int month);
 };
 
 #endif // MDATE_H